// src/app/types/movie.ts
export type Movie_API = {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};

export type Movie = Pick<
  Movie_API,
  'id' | 'title' | 'overview' | 'release_date'
> & { imageUrl: string };